helheimr/requirements.txt
libconf
python-telegram-bot
requests
emoji
psutil
pyowm
python-dateutil
pylint
rpi-rf
scipy
